feat: fs handler
